refactor(blocks-construct): taking the auth public routes in the constant file for reusable purpose
