+++ src/constant/auth-public-routes.ts
@@ -1,0 +1,11 @@
+export const publicRoutes = [
+  '/login',
+  '/signup',
+  '/sent-email',
+  '/activate',
+  '/resetpassword',
+  '/success',
+  '/activate-failed',
+  '/forgot-password',
+  '/verify-key',
+];
